finisshed base remote game

+++ 4_TicTacToe/app/src/main/java/pt/isel/pdm/tictactoe/viewmodel/ConnectViewModel.kt
@@ -4,50 +4,61 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.viewModelScope
+import com.google.protobuf.Internal.BooleanList
+import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.Job
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import pt.isel.pdm.tictactoe.model.Lobby
 import pt.isel.pdm.tictactoe.repository.UserRepository
 import pt.isel.pdm.tictactoe.services.RemoteGameService
 
 class ConnectViewModel(
-    val remoteGameService: RemoteGameService,
-    val userRepository: UserRepository
+    private val remoteGameService: RemoteGameService,
+    private val userRepository: UserRepository
 ) : ViewModelBase() {
 
     var lobbies by mutableStateOf<List<Lobby>>(emptyList())
     var currentLobby by mutableStateOf<Lobby?>(null)
+    var gameId by mutableStateOf<String?>(null)
+
+    val isJoiningOrWaitingForPlayer: Boolean
+        get() = _waitingJob != null
 
     private var _waitingJob: Job? = null
 
     fun loadLobbies() {
-        viewModelScope.launch {
+        safeViewModelScopeLaunch {
             lobbies = remoteGameService.getLobbies()
         }
     }
 
     fun createLobby() {
-        _waitingJob = viewModelScope.launch {
+        _waitingJob = safeViewModelScopeLaunch {
             val userName = userRepository.getUserData()!!.userName;
             currentLobby = remoteGameService.create(userName)
-            remoteGameService.waitForPlayer(currentLobby!!)
+            gameId = remoteGameService.waitForPlayer(currentLobby!!)
         }
     }
 
     fun joinLobby(lobby: Lobby) {
-        _waitingJob = viewModelScope.launch {
+        _waitingJob = safeViewModelScopeLaunch {
             currentLobby = lobby
-            remoteGameService.join(lobby)
+            gameId = remoteGameService.join(lobby, userRepository.getUserData()!!.userName)
         }
     }
 
     fun leaveLobby() {
-        viewModelScope.launch {
+        if (_waitingJob == null) return
+
+        safeViewModelScopeLaunch {
             val lobby = currentLobby!!
             currentLobby = null
             _waitingJob?.cancel()
+            _waitingJob = null
             remoteGameService.remove(lobby)
         }
+
     }
